+++ packages/addon-ui/ui/pages/mods.tsx
@@ -1,0 +1,9 @@
+import "./mods.css";
+
+export default function Mods() {
+	return (
+		<section>
+			<span>Work in progress.</span>
+		</section>
+	);
+}