+++ CHANGELOG.md
@@ -11,9 +11,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added the customization menu, a brand-new menu that allows you to customize almost every aspect of the fluid
   simulation and renderer.
+- **Note**: By default, you can press **M** to open the menu.
 - Adds console variables for most customization options.
-- Adds a new console command, `gelly_toggle_customization_menu`, to open and close the customization menu.
-- **Note**: Parts of the menu are WIP, you can open the menu with M or by manually binding the command.
+- Adds a new console command, **gelly_toggle_customization**, to open and close the customization menu.
+- **Note**: You may bind this to any key and Gelly will remove the M keybind for the customization menu.
+- New version popup now has bolded text.
+
+### Changed
+
+- The 'max particles' setting is now a slider in the customization menu and a console variable.
+- The range for the 'max particles' setting is now 10,000 to 1,500,000.
 
 ### Removed